Add first pass code of MVP1
The space bar currently changes the scene to the view of the door. We can switch this to a button instead, as a next step. I had trouble coding it to toggle, but I’ll leave it for now as that code would probably need to change anyway once we have a button.

Should key be sprite, not a boolean variable? Or perhaps it needs to be both? For now it’s a Boolean.

I have made the door a sprite for now, as an easy way to detect if someone has clicked on that part of the background. I wanted to make it invisible so players would just click on the door, but if it’s invisible then it won’t detect collision. I guess we need to replace it with an image of the door that exactly matches the background of the door? Not very satisfactory. Is there a better way to do it? If we keep it like this, we need to work out how to change the shape so it matches the shape of the door. (Will it scale properly if people change the size of their window?)